==> code/04-evaluate_herbaldb.py <==
from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_samples, silhouette_score
import matplotlib.pyplot as plt
import matplotlib.cm as cm
import numpy as np
import pandas
import math

# config
dataset_file = '../dataset/HerbalDB.csv'

def fillna(df):
    for col in df.columns[df.isna().any()].tolist():
        df[col].fillna(0, inplace = True)
    return df

# read dataset
df = pandas.read_csv(dataset_file)
X = fillna(df.iloc[:, 1:]).values

range_n_clusters = [5, 4, 3, 2]
n = int(math.sqrt(len(range_n_clusters)))
# Create a subplot with 1 row and 2 columns
fig, axes = plt.subplots(n, n)
fig.set_size_inches(18, 7)

for axs in axes:
    for ax1 in axs:
        n_clusters = range_n_clusters.pop()

        # The silhouette coefficient can range from -1, 1 but in this example all
        # lie within [-0.1, 1]
        ax1.set_xlim([-0.1, 1])
        # The (n_clusters+1)*10 is for inserting blank space between silhouette
        # plots of individual clusters, to demarcate them clearly.
        ax1.set_ylim([0, len(X) + (n_clusters + 1) * 10])

        # Initialize the clusterer with n_clusters value and a random generator
        # seed of 10 for reproducibility.
        clusterer = KMeans(n_clusters=n_clusters, random_state=10)
        cluster_labels = clusterer.fit_predict(X)

        # The silhouette_score gives the average value for all the samples.
        # This gives a perspective into the density and separation of the formed
        # clusters
        silhouette_avg = silhouette_score(X, cluster_labels)
        print("For n_clusters =", n_clusters, "The average silhouette_score is :", silhouette_avg)

        # Compute the silhouette scores for each sample
        sample_silhouette_values = silhouette_samples(X, cluster_labels)

        y_lower = 10
        for i in range(n_clusters):
            # Aggregate the silhouette scores for samples belonging to
            # cluster i, and sort them
            ith_cluster_silhouette_values = \
                sample_silhouette_values[cluster_labels == i]

            ith_cluster_silhouette_values.sort()

            size_cluster_i = ith_cluster_silhouette_values.shape[0]
            y_upper = y_lower + size_cluster_i

            color = cm.nipy_spectral(float(i) / n_clusters)
            ax1.fill_betweenx(np.arange(y_lower, y_upper),
                            0, ith_cluster_silhouette_values,
                            facecolor=color, edgecolor=color, alpha=0.7)

            # Label the silhouette plots with their cluster numbers at the middle
            ax1.text(-0.05, y_lower + 0.5 * size_cluster_i, str(i))

            # Compute the new y_lower for next plot
            y_lower = y_upper + 10  # 10 for the 0 samples

        ax1.set_title("The silhouette plot for n_cluster = {} with avg {:.2f}".format(n_clusters, silhouette_avg), fontsize=10)
        ax1.set_xlabel("The silhouette coefficient values", fontsize=8)
        ax1.set_ylabel("Cluster label", fontsize=8)

        # The vertical line for average silhouette score of all the values
        ax1.axvline(x=silhouette_avg, color="red", linestyle="--")

        ax1.set_yticks([])  # Clear the yaxis labels / ticks
        ax1.set_xticks([-0.1, 0, 0.2, 0.4, 0.6, 0.8, 1])

        # # 2nd Plot showing the actual clusters formed
        # feature_index = (520, 401)
        # colors = cm.nipy_spectral(cluster_labels.astype(float) / n_clusters)
        # ax2.scatter(X[:, feature_index[0]], X[:, feature_index[1]], marker='.', s=30, lw=0, alpha=0.7, c=colors, edgecolor='k')

        # # Labeling the clusters
        # centers = clusterer.cluster_centers_

        # # Draw white circles at cluster centers
        # ax2.scatter(
        #     centers[:, feature_index[0]], 
        #     centers[:, feature_index[1]], 
        #     marker='o',
        #     c="white", 
        #     alpha=1, 
        #     s=200, 
        #     edgecolor='k'
        # )

        # for i, c in enumerate(centers):
        #     ax2.scatter(
        #         c[feature_index[0]], 
        #         c[feature_index[1]], 
        #         marker='$%d$' % i, 
        #         alpha=1,
        #         s=50, 
        #         edgecolor='k'
        #     )

        # ax2.set_title("The visualization of the clustered data.")
        # ax2.set_xlabel("Feature space for the 1st feature")
        # ax2.set_ylabel("Feature space for the 2nd feature")

plt.suptitle("Silhouette analysis for HerbalDB dataset")      
plt.tight_layout()
plt.subplots_adjust(top=0.9)
plt.show()
